+++ src/runner.ts
@@ -1,7 +1,7 @@
 import * as core from "@actions/core";
 import { context, getOctokit } from "@actions/github";
 import { getPRsLabels, getWorkflowRunJobs } from "./github/github";
-import { traceWorkflowRunJobs } from "./tracing/job";
+import { traceWorkflowRun } from "./tracing/job";
 import { type Attributes, createTracerProvider } from "./tracing/trace";
 
 async function run() {
@@ -12,14 +12,14 @@ async function run() {
   const ghToken = core.getInput("githubToken") || process.env["GITHUB_TOKEN"] || "";
   const octokit = getOctokit(ghToken);
 
-  core.info(`Get Workflow Run Jobs for ${runId}`);
+  core.info(`Get workflow run jobs for ${runId}`);
   const workflowRunJobs = await getWorkflowRunJobs(context, octokit, runId);
 
   core.info("Get PRs labels");
   const prNumbers = workflowRunJobs.workflowRun.pull_requests?.map((pr) => pr.number) ?? [];
   const prLabels = await getPRsLabels(context, octokit, prNumbers);
 
-  core.info(`Create Trace Provider for ${otlpEndpoint}`);
+  core.info(`Create tracer provider for ${otlpEndpoint}`);
   const attributes: Attributes = {
     serviceName: otelServiceName || workflowRunJobs.workflowRun.name || `${workflowRunJobs.workflowRun.workflow_id}`,
     serviceVersion: workflowRunJobs.workflowRun.head_sha,
@@ -34,11 +34,11 @@ async function run() {
   const provider = createTracerProvider(otlpEndpoint, otlpHeaders, attributes);
 
   try {
-    core.info(`Trace Workflow Run Jobs for ${runId} and export to ${otlpEndpoint}`);
-    const traceId = await traceWorkflowRunJobs(workflowRunJobs, prLabels);
+    core.info(`Trace workflow run for ${runId} and export to ${otlpEndpoint}`);
+    const traceId = await traceWorkflowRun(workflowRunJobs, prLabels);
     core.setOutput("traceId", traceId);
 
-    core.info("Flush and shutdown trace provider");
+    core.info("Flush and shutdown tracer provider");
     await provider.forceFlush();
     await provider.shutdown();
     core.info("Provider shutdown");